refactor(api): move checkCampaignParticipationBelongsToUser in participation context
